fix: Load all extraDiagrams
