+++ packages/mermaid/src/config.type.ts
@@ -3,7 +3,7 @@
 import DOMPurify from 'dompurify';
 
 export interface MermaidConfig {
-  extraDiagrams?: any;
+  extraDiagrams?: string[];
   theme?: string;
   themeVariables?: any;
   themeCSS?: string;
