+++ packages/mermaid/src/diagrams/class/classDetector.ts
@@ -1,4 +1,4 @@
-import type { DiagramDetector } from '../../diagram-api/detectType';
+import type { DiagramDetector } from '../../diagram-api/types';
 
 export const classDetector: DiagramDetector = (txt, config) => {
   // If we have confgured to use dagre-wrapper then we should never return true in this function
