+++ packages/mermaid/src/diagrams/er/erDetector.ts
@@ -1,4 +1,4 @@
-import type { DiagramDetector } from '../../diagram-api/detectType';
+import type { DiagramDetector } from '../../diagram-api/types';
 
 export const erDetector: DiagramDetector = (txt) => {
   return txt.match(/^\s*erDiagram/) !== null;
