+++ packages/mermaid/src/diagrams/gantt/ganttDetector.ts
@@ -1,4 +1,4 @@
-import type { DiagramDetector } from '../../diagram-api/detectType';
+import type { DiagramDetector } from '../../diagram-api/types';
 
 export const ganttDetector: DiagramDetector = (txt) => {
   return txt.match(/^\s*gantt/) !== null;
