+++ packages/mermaid/src/diagrams/git/gitGraphDetector.ts
@@ -1,4 +1,4 @@
-import type { DiagramDetector } from '../../diagram-api/detectType';
+import type { DiagramDetector } from '../../diagram-api/types';
 
 export const gitGraphDetector: DiagramDetector = (txt) => {
   return txt.match(/^\s*gitGraph/) !== null;
