+++ packages/mermaid/src/diagrams/info/infoDetector.ts
@@ -1,4 +1,4 @@
-import type { DiagramDetector } from '../../diagram-api/detectType';
+import type { DiagramDetector } from '../../diagram-api/types';
 
 export const infoDetector: DiagramDetector = (txt) => {
   return txt.match(/^\s*info/) !== null;
