+++ packages/mermaid/src/diagrams/sequence/sequenceDetector.ts
@@ -1,4 +1,4 @@
-import type { DiagramDetector } from '../../diagram-api/detectType';
+import type { DiagramDetector } from '../../diagram-api/types';
 
 export const sequenceDetector: DiagramDetector = (txt) => {
   return txt.match(/^\s*sequenceDiagram/) !== null;
