+++ packages/mermaid/src/diagrams/user-journey/journeyDetector.ts
@@ -1,4 +1,4 @@
-import type { DiagramDetector } from '../../diagram-api/detectType';
+import type { DiagramDetector } from '../../diagram-api/types';
 
 export const journeyDetector: DiagramDetector = (txt) => {
   return txt.match(/^\s*journey/) !== null;
